Created fcnClearLogs Function
Created fcnClearLogs function to the file. When this function runs, it will check to see if certain files in the Temp directory have been in the directory for at least two days. At the start of a new month or new year, the deletion period may extend up to four days. Only CSV files that start with "CADAttInLog - " will be evaluated. This function is meant to be triggered only at the startup of the CAD application.